카드 이미지 수정중

+++ src/main/java/com/example/week08/service/PlaceService.java
@@ -8,6 +8,7 @@ import com.example.week08.dto.request.PlacePutDto;
 import com.example.week08.dto.request.PlaceRequestDto;
 import com.example.week08.dto.request.PostRequestDto;
 import com.example.week08.errorhandler.BusinessException;
+import com.example.week08.errorhandler.ErrorCode;
 import com.example.week08.repository.PlaceRepository;
 import com.example.week08.repository.PostRepository;
 import com.example.week08.util.S3Uploader;
@@ -31,61 +32,53 @@ public class PlaceService {
 
     //카드 생성
     @Transactional
-    public void placeCreate(Long courseId, PlaceRequestDto placeRequestDto
-//            , Member member
-    )
-//            throws IOException
+    public void placeCreate(Long courseId, PlaceRequestDto placeRequestDto, MultipartFile image)
+            throws IOException
     {
-//        String imageDefault = "https://dh-s3-bucket01.s3.ap-northeast-2.amazonaws.com/%EB%94%94%ED%8F%B4%ED%8A%B8%EC%9D%B4%EB%AF%B8%EC%A7%80.png";
-
+        String placeImage = s3Uploader.upload(image, "static");
         Post post = postRepository.findById(courseId).orElseThrow(
                 () -> new BusinessException("추천 코스가 존재하지 않습니다.", POST_NOT_EXIST)
         );
 
-//        if (placeRequestDto.getImage() == null){
-//            placeRequestDto.setImage(imageDefault);
-//        }
-//        String placeImage = s3Uploader.upload(image, "static");
-
-        Place place = new Place(post, placeRequestDto
-//                , member
-        );
+        Place place = new Place(post, placeRequestDto, placeImage);
         placeRepository.save(place);
     }
 
     //카드 삭제
     @Transactional
-    public void placeDelete( Long placeId
-//            , Member member
-    ) {
+    public void placeDelete( Long placeId, Member member) {
         Place place = placeRepository.findById(placeId).orElseThrow(() ->
-                new BusinessException("카드가 존재하지 않습니다.", PLACE_NOT_EXIST)
+                new BusinessException("카드가 존재하지 않습니다.", ErrorCode.PLACE_NOT_EXIST)
         );
-//        Post post = postRepository.findById(courseId).orElseThrow(
-//                () -> new BusinessException("추천 코스가 존재하지 않습니다.", POST_NOT_EXIST)
-//        );
-//        if (!place.getMember().getId().equals(member.getId())) {
-//            throw new IllegalArgumentException("수정 권한이 없습니다.");
-//        }
+
+        if (!place.getMember().getId().equals(member.getId())) {
+            throw new IllegalArgumentException("삭제 권한이 없습니다.");
+        }
+
         placeRepository.deleteById(placeId);
     }
 
     //카드 수정
     @Transactional
-    public void placeUpdate(Long courseId, PlacePutDto placePutDto
-//            , Member member
-    ) {
+    public void placeUpdate(Long courseId, PlacePutDto placePutDto, MultipartFile image, Member member)throws IOException {
         Place place = placeRepository.findById(placePutDto.getPlaceId()).orElseThrow(() ->
                 new BusinessException("카드가 존재하지 않습니다.", PLACE_NOT_EXIST)
         );
         Post post = postRepository.findById(courseId).orElseThrow(
                 () -> new BusinessException("추천 코스가 존재하지 않습니다.", POST_NOT_EXIST)
         );
-//        if (!place.getMember().getId().equals(member.getId())) {
-//            throw new IllegalArgumentException("수정 권한이 없습니다.");
-//        }
+        if (!place.getMember().getId().equals(member.getId())) {
+            throw new IllegalArgumentException("수정 권한이 없습니다.");
+        }
+        String imageUrl = post.getImage();
+        //이미지 존재시 먼저 삭제후 다시 업로드.
+        if (imageUrl != null) {
+            String deleteUrl = imageUrl.substring(imageUrl.indexOf("static"));
+            s3Uploader.deleteImage(deleteUrl);
+            imageUrl = s3Uploader.upload(image, "static");
+        }
 
-        place.update(place, placePutDto, post);
+        place.update(place, placePutDto, post, imageUrl);
     }
 
 }
